+++ backend/plugins/jira/tasks/issue_convertor.go
@@ -103,7 +103,7 @@ func ConvertIssues(taskCtx plugin.SubTaskContext) errors.Error {
 				Priority:                jiraIssue.PriorityName,
 				CreatedDate:             &jiraIssue.Created,
 				UpdatedDate:             &jiraIssue.Updated,
-				LeadTimeMinutes:         int64(jiraIssue.LeadTimeMinutes),
+				LeadTimeMinutes:         jiraIssue.LeadTimeMinutes,
 				TimeSpentMinutes:        jiraIssue.SpentMinutes,
 				OriginalProject:         jiraIssue.ProjectName,
 				Component:               jiraIssue.Components,
